// src/app/models/users.ts
export interface UserDto {
    usersPkId: string;
    usersStPassword: string;
    
    usersStNom: string;
    usersStPrenom: string;
    usersStMatricule: string;
    
    usersStStructure: string;
    usersStFonction: string;
    usersStCodeStructure: string;
    usersInEtatCompte: number;
    usersStImageName: string;
    usersStProfile: string;
}